content/base/test/test_bug419132.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=419132
-->
<head>
  <title>Test for Bug 419132</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=419132">Mozilla Bug 419132</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<iframe src="bug419132.html"></iframe>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 419132 **/

SimpleTest.waitForExplicitFinish();

setTimeout(finish, 60000)

function finish()
{
  ok(true, "This is a mochikit version of a crash test. To complete is to pass.");
  SimpleTest.finish();
}


</script>
</pre>
</body>
</html>
